pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "server-client-test"
version = "0.1.0"
description = "Policy server/client communication test toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "PyYAML",
  "h5py",
  "msgpack",
  "websockets",
  "opencv-python",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
